requirements.txt
streamlit
pandas
plotly
altair
matplotlib
seaborn
numpy
pillow
